+++ src/pdf_genesis/render/__init__.py
@@ -1,11 +1,10 @@
 from pdf_genesis.render.bench import render_bench_pdf
-from pdf_genesis.render.chorus import render_chorus_pdf
 from pdf_genesis.render.design import render_design_pdf
-from pdf_genesis.render.patent import render_patent_pdf
+from pdf_genesis.render.physics import render_chorus_pdf, render_physics_pdf
 
 __all__ = [
+    "render_physics_pdf",
     "render_chorus_pdf",
     "render_design_pdf",
-    "render_patent_pdf",
     "render_bench_pdf",
 ]